+++ R/Animate.R
@@ -1,0 +1,17 @@
+#' Generic Animate Method
+#' 
+#' Animate graph-plot objects from select qdap outputs.
+#' 
+#' @param x A qdap object (list) with a animatable graph-plot object (e.g., 
+#' \code{\link[qdap]{discourse_map}}).
+#' @param \ldots Arguments passed to Animate method of other classes.
+#' @export
+#' @seealso \code{\link[qdap]{scores}},
+#' \code{\link[qdap]{counts}},
+#' \code{\link[qdap]{preprocessed}},
+#' \code{\link[qdap]{proportions}}
+#' @return Returns a plot object.
+Animate <-
+function(x, ...){
+    UseMethod("Animate")
+}